requirements.txt
openai
torch
transformers
Pillow
gTTS
Flask
